BREAKING: plugin-configs now must have IDs equal to their plugins names
Previously, you configured a plugin by making a plugin-config resource with attributes.module equal to the plugin's name. By convention, we often also made the ID of the plugin-config equal to the plugin's name, but that was only convention.

This change makes the use of the plugin name as the plugin-config's ID formally required and removed attributes.module.

There was no good reason to keep these things distinct, it's an error to have multiple plugin configs for the same plugin anyway. Plus our ability to understand the implicit dependency between things like data-sources and their corresopnding plugin-configs was hampered by this distinction.