Add handshake to ember-cli connection to work around Docker weirdness
If a container has a published port, but there's no actual listener
bound to the port inside the container, Docker does some weird stuff.
Attempting to connect to a port with no listeners should refuse the
connection, but docker accepts it, then immediately closes it. So, if
our connection is closed, just try again until the handshake succeeds.