drop ember-cli proxy
Previously, in both dev and test we proxied requests from the ember app to the hub through ember-cli. Mostly because we used to always run the hub inside ember-cli, and proxying was an easy transition step.

This change removes the proxy and makes all requests go directly to the hub, bypassing ember-cli.

In addition to being simpler, this has the benefit of forcing the app to use CORS in dev and test, so you won't get any surprises if your production environment requires CORS.