better port control
This adds `--port` and `--url` options to `ember hub:start`. It needs `--url` in production so that the ember app's it helps build via codegen will know where to find it.

I also added an optional way to alter the chosen port if you are letting the hub start automatically via `ember s` or `ember test`: set the `HUB_PORT` environment variable.