Switch indexers.update() from realTime to forceRefresh
This makes us _always_ wait for content to be visible before resolving the update() promise. That is basically always what you want (if you don't care about it, you can still just not await the promise).

Before, we would choose between forcing an immediate hard refresh before resolving, or we would just resolve immediately without waiting for content to be visible in the search index.

Now, we either wait for the next scheduled refresh to happen, or we force a refresh to happen, but in either case when `update()` resolves you can be confident that your content is visible in the search index.