unify client & server jsonapi-factory modules
We had two copies of this module because one was an ECMA module for the client build and one was CJS for node. Which is obviously not great.

And there was even a bugfix in one but not the other.

This unifies them, taking advance of `@std/esm` for ECMA module interop within node.

I'm doing the interop all local to the test-support plugin, but in the future we may want to blanket enable it for the whole Hub and start authoring anything in modules.