Sort record attributes deterministically before save #692
This PR has an effect on previously persisted data.

Before, when you saved a record with the git writer, the resulting
stringified JSON could have a meaningless diff, where the attributes
drifted around. This is because JSON.stringify is not deterministic.
This PR uses a library to deterministically sort keys before
saving the JSON. This change will result in a major diff the first
time past data is modified, but smaller, meaningful diffs afterwards.

Closes #692