make eslint respect gitignore
This switches eslint to ignore the same things that are in gitignore, which allows us to keep our list of already-converted-to-typescript files in only one place.

As a side effect of unifying eslintignore and gitignore, this enabled linsting of the many test addons under the tests dir, which required fixing some issues with missing node engine declarations and dependencies.