Add image thumbs
After looking into imgix for static deployments, I figured out a way we could
generate signed urls on build. It requires patches to both broccoli-asset-rev
and broccoli-fingerprint.

Instead, I made a quick script to generate thumbs with imagemagick for static
deploy to solve the immediate performance issue.

It also adds the urls into the JSON files automatically.

I think imgix is a good choice for an app with a backend server that can
sign stuff, but not for static site deployment.