Add prepaid card issued confirmation card (#1681)
* Add preliminary display of confirmed prepaid card

This only shows the header background and image for now.

* Add issuer name to confirmation card

* Add placeholders for other card values

This is duplicated in …LayoutCustomization::DisplayView and
…::FormView but it’s probably fine to use here until the face
value-choosing step is implemented.

* Add card container wrapper

This isn’t an action so maybe this is the incorrect wrapper,
but other things I found lacked a `::Section` 🤔

* Add confetti background

This includes changing the background of CardPay::PrepaidCard
to white, otherwise a transparent-background card header shows
the confetti. It also adds the :where pseudo-class to
.prepaid-card to allow its box shadow to be overwritten,
thanks to @Aierie for introducing me to that.

* Add test selector to assert on card issuer name

* Change assertion to use parent selector

This seems preferable to asserting that the card exists twice.