Update call to actually use chosen value
This is a bit of a convoluted test interface, I just find it
unsatistfying that I could change the actual call to create
a prepaid card without having to change the tests at all.
What’s ensuring that the chosen value is actually threading
through? I’d usually be using Sinon or the like for this 😯