web: Add merchant info repersistance after change (#2029)
This addresses a bug where if a user clicked the button to create a merchant
but then rejected the request in Card Wallet, if they then changed their
merchant details and tried again, the old details would be attached to the
merchant safe. This updates the merchant customisation card to delete
a persisted merchant DID from the session if the details have changed
so a new one will be persisted in the following card.

Dealing with the orphaned DID is out of scope for this PR.

This might produce confusion if the user goes back to edit their details but
leaves the slug the same, as it will then be invalid because the old details
have already been persisted.