Add fix for failing test
@Aierie found this fix; we paired for a while and never fully
understood the nature of it, but it’s somewhat of a race
condition in HubAuthentication’s initializeTask and
hasValidAuthentication. Changing the way storeWithdrawalLimits
is hooked on to the end of Layer2Service.strategy.initializeTask
fixes the problem, which suggests it’s something about the use
of then on the task instance, but we didn’t get full clarity
on that. Ultimately it would be nice to have an interface
for registering callbacks to occur after layer 2 has
been initialised, as this kind of error could still arise.