web: Add validation enforcement of withdrawal limits (#2091)
This adds checks that the withdrawal amount is within the limits
returned by TokenBridgeHomeSide.getWithdrawalLimits and includes the
relevant limit in the validation string if not.

This necessitated some adjustments to the validation function, which had
a min option that was being used in the deposit transaction amount card
to enforce that the value had to be higher than 0, which isn’t a true
minimum. min is now optional and > 0 is enforced in its absence. This
didn’t have test representation so I added another test for that card
and refactored out the duplicated setup.

The validation function now also accepts a balance to distinguish
between the maximum from the withdrawal limit vs the maximum from the
balance.

We decided to not add the bridged tokens to the SDK’s native-currencies
list, so there’s some string manipulation with the result of the call to
convertRawAmountToNativeDisplay to display a bridged token in the
validation message if that’s the context.

The change to the persisted issue prepaid card workflow test indicates
a race condition that should eventually be addressed, this is a workaround.