drop dependency on `esm`
The `esm` package is a polyfill for ES modules in node. It is unmaintained now that node has native support. Node's native support however is stricter and hard to adopt. Our best immediate path forward is to allow typescript to emit commonjs instead of ES modules.

This was all motivated by the need to support Node 14+ features that allow customizing the `exports` key in package.json. `esm` does not implemenet this which breaks the card compiler.