web: Change L2 connect to always persist address (#2163)
This fixes a subtle discrepancy where workflow restoration
would abort when the layer 2 account had changed when it had
originally been started without a wallet connected. If the
workflow had been started with a layer 2 account connected,
restoring it would not abort even if the account had changed.

Now the layer 2 address will always be persisted alongside the
other workflow data so it can be compared during restoration.

This also removes an outdated test placeholder comment.