Fix conflicting SVG part IDs
ember-svg-jar was minifying the id attributes within our SVGs to the point that they conflict with the IDs in other SVGs when they are both inlined into the page at the same time. This PR to disables minifying those IDs and gives unique IDs to the SVGs that have this issue (likely from having the same damage done to them by ImageOptim).