web: Add SafeBalanceCard
This is an extraction from #2241 that @Aierie collaborated on.
CardPay::SafeBalanceCard can be used at the end of a workflow to show
the balances of the safe that was part of a transaction. It’s used here
in withdrawal, which previously would only show the depot balance
(LayerTwoConnectCard) even if a merchant safe had been used.

Since workflow cards can’t take parameters, this instead expects
configuration via a safeBalanceCardKey value in the workflow session
and throws errors if it isn’t found. For example, safeBalanceCardKey
for the withdrawal workflow is withdrawalSafe.