web: Fix merchant source issuance problems (#2241)
* Update BalanceViewBanner to handle merchant safes

* web: Add SafeBalanceCard

This is an extraction from #2241 that @Aierie collaborated on.
CardPay::SafeBalanceCard can be used at the end of a workflow to show
the balances of the safe that was part of a transaction. It’s used here
in withdrawal, which previously would only show the depot balance
(LayerTwoConnectCard) even if a merchant safe had been used.

Since workflow cards can’t take parameters, this instead expects
configuration via a safeBalanceCardKey value in the workflow session
and throws errors if it isn’t found. For example, safeBalanceCardKey
for the withdrawal workflow is withdrawalSafe.

* Add filtering of safes without sufficient balance

* Change to not default to low-balance depot safe

There’s some possible weirdness here… need to get
more clarification.

* Add merchant options to pre-issuance balance check

* Fix test issuance to use source vs hardcoded depot

* Fix bug where depot safe doesn’t exist

* Change wording of insufficient funds message

It seems subpar to have this duplication in the tests…?

* Change issuance workflow to end with safe balance

This relies on the work in #2254, which awaits review.

* Move issuance source filtering into resource

There are several avenues for improvement here. Ultimately
should the safes resource be in Layer2Network so
daiMinValue can live there? etc

* Fix update-after-use warning

This fixes the below:
You attempted to update `selectedSafe` on `FundingSourceCard`,
but it had already been used previously in the same computation.

* Add handling for post-restore insufficient balance

It would be preferable for the workflow’s check to run
to stop restoration from proceeding if there is no
sufficient balance but this will do in the meantime.

* Remove unused getter

* Update with new SafeBalanceCard config setup

* Update factory insertion for new sorting

* Add direct test for issuance source filtering

This is already exercised elsewhere but I like having something
testing exactly what’s returned from this kind of filtering.

* Remove unused function

This has since been fixed in main.

* Change variable name

* Update how minima are stored and accessed

* Add error to warn when minima haven’t loaded

* Add restoration cancelation for insufficient safe

* fetch prepaid card issuance min amounts in strategy (#2299)

* remove fixme

* rename minimumFaceValue -> minimumDaiValue

Co-authored-by: Aierie <39579264+Aierie@users.noreply.github.com>
Co-authored-by: Aierie <michael.khor@cardstack.com>