web: Add merchant safes as issuance source options (#2208)
This follows a pattern similar to #2038. Cards in the issuance workflow
had hard-coded assumptions that the depot safe would be the source
for funding the prepaid card, so this adds the SafeChooser and persists
prepaidFundingSafeAddress to the workflow.

The SafeBalanceCard replaces the LayerTwoConnect card at the end
of the workflow to show updated balances for the relevant safe.