web: Rename merchants to business accounts
This is a user-facing change only, since the SDK will continue
to call it a merchant safe, and it seemed too dramatic to
change Hub routes and database names. It replaces mentions of
“merchant” with “business” and “business account” depending
on what makes the most sense. It also renames the route from
merchant-services to business-services and renames the flow
query parameter from create-merchant to create-business.