web: Change cancel to resend for deposit workflow (#2393)
This modifies the deposit workflow to use a retry UI that allows us to
keep track of every deposit signing request sent until one returns with
a transaction hash, and then we continue following only that one. The
reason for this is that we don't reliably have a way to cancel
transactions via nonce.