Add preliminary tests for NotifyMerchantClaim task
I changed the task to use the injected SDK to facilitate
mocking. I think this general approach should be workable
for the other notification task and hopefully in general.

I don’t love the need to attach a then but it seems to be
the way: https://www.chaijs.com/plugins/chai-as-promised/