hub: Add stubbed notifications (#2434)
This adds a ContractSubscriptionEventHandler that starts within the Hub
worker. It subscribes to contract events (CustomerPayment and
MerchantClaim hardcoded for now) and creates worker jobs when they are
received. Those jobs extract the relevant data from the events and
create `send-notifications` jobs with prototype push notifications to
display.

The `send-notifications` task just prints the prototype notification to
logs for now, pending the ability to actually send push notifications.