hub: Change incidents to have shorter names (#2605)
This shifts what was the name to the body when creating a Statuspage
incident to allow the name to be a shorter string (“Transactions
delayed” only, for now). When you create a new incident, body is the
“initial message, created as the first incident update.”

The desire is to have a more concise message to show in Card Wallet:

Since web client’s design is more able to facilitate longer messages, it
uses the name as a prefix followed by the body of the last
incident_updates entry.