Make install script work on both OSX and Linux
mktemp works differently on osx and linux.
The -t flag specifies that a template should be used, and needs
to contain XXXX characters for the random part.

On OSX it is a flag for specifying a prefix, and the template
comes later.

Using it with no arguments should work correctly on both platforms.