hub: Change spaces to belong to merchants (#2595)
This sets up a has_one-type relationship between merchant_infos and
card_spaces, where a card_spaces cannot exist without linking to a
merchant_infos. It also removes the url and ownerAddress fields in
card_spaces. merchant_infos.slug is the unique identifier that will be
used in #2586 to fetch the details of a Card Space.