Clean up image uploading and editing (#2669)
This prepares image uploading components for usage in and out of workflows by separating UI state/trigger concerns from the logic of uploading and editing + hiding the messy and fragile state management of image editing. The image uploading functionality is handled by common/image-upload-action and the UI is handled by common/image-upload-interface.