sdk/web: Restore withdrawal with dynamic estimation (#2877)
This is a reversal of the web client changes in #2875 and an SDK change
to dynamically estimate withdrawal gas. When run with this, the gas
limit calculated was 4k higher than the previously hardcoded gas value
of 350k.

This leaves another use of the hardcoded value intact in
getEstimatedGasForWithdrawalClaim, which is used at the beginning of the
withdrawal workflow to ensure the user has enough in their wallet to
cover gas. This should be updated when possible.