hub: Add POST /api/email-card-drop-requests (#2902)
This adds an endpoint to start the email card drop process. A successful
request triggers stubbed jobs to send the verification code and
subscribe the email to a list.

It includes validation: the user must be authenticated with Hub, isEmail
from the validator package must be true, and the email and EOA must not
have already received a prepaid card.