hub: Change claim check to use all claimed_at (#2917)
This changes GET /api/email-card-drop-requests to detect the presence of
claimed_at in the returned rows instead of transaction_hash, since the
latter will not be present until the drop-card task has run. It also
checks all returned rows instead of just the first one to ensure that a
claim any email associated with EOA is detected.