hub: Add reservation/quantity check for drop requests (#2974)
This adds a 503 response from POST /api/email-card-drop-requests if
SDK’s PrepaidCardMarketV2.getQuantity ≤ the number of “active”
reservations, as in new records in the table from the past 60 minutes
that haven’t been claimed. It also 503s if the contract is paused.

There’s a Sentry alert if getQuantity - reservations < 25 (in config).