wc: Add subscription override to remove parameters (#2996)
This is a patch to eth-block-tracker`s SubscribeBlockTracker to fix bugs
relating to web-client missing blockchain events, which has resulted in
symptoms like withdrawal never completing and needing to be refreshed
and the depot balances not updating after a deposit and needing to be
refreshed. The patch exists simply to remove this line:
https://github.com/MetaMask/eth-block-tracker/blob/e7c075a80e46c18ecbd44d007f2257ba31079c8b/src/SubscribeBlockTracker.ts

Some RPC node implementations reject eth_subscribe to newHeads with an
empty parameter object but they all appear to support it without a
parameter object, so this removes it entirely.