hub: Fix card drop availability calculation (#3028)
A combination of discrepancy in mock response types and test setup
coincidences were masking a bug in the determination of whether there
are sufficient funds for a user to initiate the card drop process. To
exercise #3025 in staging, I checked getQuantity (48) and added 100
email_card_drop_requests records to simulate the number of reserved
cards exceeding available funds and found that it was still showing
available: true.

With this change, available: false was the result in the same
circumstances.