hub: Add CreateProfile task (#3016)
This adds the CreateProfile task that’s part of the in-app purchase of a
merchant profile. It will be triggered by a POST /api/profile-purchases,
which will be responsible for creating the merchant_infos record, so it
receives merchant-info-id and job-ticket in its payload.

When everything goes as planned, the task:

* calls the relay server endpoint to register a profile
* waits for the transaction and then queries the subgraph for the
  newly-created merchant safe address
* updates the job_tickets record with the safe address as a result
* queues the persist-off-chain-merchant-info task

It was necessary to add waitForTransactionConsistency as a new
cardpay-sdk export.