Don't rely on the issuer in CreatePrepaidCard being correct
Currently, the contract is sending the wrong issuer (it's actually the owner) in the event. To avoid implementing versioning, after we
fix the event params sent by the contract, we get the issuer and the owner from the source of truth. This introduces another
query to the contract (prepaidCardMgr.getPrepaidCardOwner), but I am not sure how this will affect the performance of this mapping.