hub: Replace card_spaces/merchant_infos with profiles (#3120)
This is mostly a mechanical process of combining the creation of
separate models into a single profile. Since the surviving fields on
card_spaces were all optional and it had a foreign key for merchants,
it’s as if it merged into merchant_infos.

This preserves most existing endpoints at /api/card-spaces and
/api/merchant-infos so we don’t have to update all clients
simultaneously. Then we can add /api/profiles and eventually remove
those once everything has migrated. POST /api/card-spaces is removed
because it’s no longer applicable.