ssr: Fix Fastboot development
There’s a bug in development with Fastboot where the attempt
to load ember-get-config in ember-cli-mirage returns undefined.
This removes Mirage’s Ember Data code, which is what is trying
to read from ember-get-config.

It also adds a deprecation exception for owner.inject, which
is being used in a Fastboot initialiser.