hub: Add optional invalidation upon write to S3 (#3157)
This updates PATCH /api/profiles/:id to trigger Cloudfront invalidation
when it writes the updated profile JSON document to S3. Without this,
updates would not be detected by clients until the cache expired, which
could take up to 24h.

The relevant Cloudfront distribution id is accessed as a secret as a
convenience for sharing it between environments, it’s not actually a
secret.