Remove superfluous component helpers
These were producing errors like that seen below. It only
happens when a component is used without arguments.

addon/components/boxel/card-picker/index.gts:59:44 - error TS2769: The type of this expression doesn't appear to be a valid value to pass the {{component}} helper. If possible, you may need to give the expression a narrower type, for example `'component-a' | 'component-b'` rather than `string`.
  No overload matches this call.
    Overload 1 of 6, '(component: never): never', gave the following error.
      Argument of type 'TemplateOnlyComponent<Signature>' is not assignable to parameter of type 'never'.
    Overload 2 of 6, '(component: null | undefined): null', gave the following error.
      Argument of type 'TemplateOnlyComponent<Signature>' is not assignable to parameter of type 'null | undefined'.